Move xmlUtils.h to src
