+++ FWCore/MessageLogger/src/xmlUtils.h
